requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
bcrypt>=4.0.0
structlog>=24.0.0
aiofiles>=23.0.0
psycopg2-binary>=2.9.0
pydantic[email]
email-validator